Add DebugConsoleWriter, Add AchievementLogicBase and IAchievementGenerator, Update AchievementManager for definition generators.
The debug console writer is a shorthand console writer that writes only when debugging is allowed in the mod settings.
The class AchievementLogicBase and SaveableAchievementLogicBase abstract the achievement base logic in a class that can easily be extended for specific achievement functionality.
The IAchievementGenerator allows to load achievement definitions without having to patch methods.